+++ django-backend/fecfiler/memo_text/test_models.py
@@ -2,7 +2,7 @@ import uuid
 from django.test import TestCase
 from .models import MemoText
 from fecfiler.committee_accounts.models import CommitteeAccount
-from fecfiler.committee_accounts.views import create_committee_view
+from fecfiler.committee_accounts.utils import create_committee_view
 from fecfiler.reports.tests.utils import create_form3x
