+++ django-backend/fecfiler/openfec/views.py
@@ -1,10 +1,14 @@
 from rest_framework import viewsets
-from django.http.response import HttpResponse
+from django.http.response import HttpResponse, HttpResponseBadRequest
 from rest_framework.response import Response
 from rest_framework.decorators import action
 from fecfiler.mock_openfec.mock_endpoints import query_filings, committee
 import requests
 import fecfiler.settings as settings
+from fecfiler.committee_accounts.utils import (
+    check_can_create_committee_account,
+    retrieve_recent_f1,
+)
 
 import structlog
 
@@ -14,6 +18,11 @@ logger = structlog.get_logger(__name__)
 class OpenfecViewSet(viewsets.GenericViewSet):
     @action(detail=True)
     def committee(self, request, pk=None):
+        check_can_create = request.query_params.get("check_can_create")
+        if check_can_create == "true" and not check_can_create_committee_account(
+            pk, request.user
+        ):
+            return HttpResponseBadRequest()
         response = committee(pk)
         if response:
             return Response(response)
@@ -44,26 +53,3 @@ class OpenfecViewSet(viewsets.GenericViewSet):
             fec_response = requests.get(f"{settings.FEC_API}filings/", params)
             response = fec_response.json()
         return Response(response)
-
-
-def retrieve_recent_f1(committee_id):
-    """Gets the most recent F1 filing
-    First checks the realtime enpdpoint for a recent F1 filing.  If none is found,
-    a request is made to a different endpoint that is updated nightly.
-    The realtime endpoint will have more recent filings, but does not provide
-    filings older than 6 months. The other endpoint returns a committee's current data
-    informed by their most recent F1 and F2 filings."""
-    headers = {"Content-Type": "application/json"}
-    params = {
-        "api_key": settings.FEC_API_KEY,
-        "committee_id": committee_id,
-    }
-    endpoints = [
-        f"{settings.FEC_API}efile/form1/",
-        f"{settings.FEC_API}committee/{committee_id}/",
-    ]
-    for endpoint in endpoints:
-        response = requests.get(endpoint, headers=headers, params=params).json()
-        results = response["results"]
-        if len(results) > 0:
-            return results[0]
