+++ django-backend/fecfiler/reports/form_3x/tests/test_models.py
@@ -1,7 +1,7 @@
 from django.test import TestCase
 from ..models import Form3X
 from fecfiler.committee_accounts.models import CommitteeAccount
-from fecfiler.committee_accounts.views import create_committee_view
+from fecfiler.committee_accounts.utils import create_committee_view
 from fecfiler.reports.tests.utils import create_form3x
