+++ django-backend/fecfiler/reports/tests/test_models.py
@@ -2,7 +2,7 @@ from django.test import TestCase
 from fecfiler.web_services.models import UploadSubmission
 from fecfiler.reports.models import Report, Form24, Form3X
 from fecfiler.reports.tests.utils import create_form3x, create_form24
-from fecfiler.committee_accounts.views import create_committee_view
+from fecfiler.committee_accounts.utils import create_committee_view
 from fecfiler.committee_accounts.models import CommitteeAccount
 from fecfiler.transactions.tests.utils import create_ie
 from fecfiler.contacts.models import Contact
