+++ django-backend/fecfiler/transactions/schedule_c/tests/test_views.py
@@ -4,7 +4,7 @@ from fecfiler.transactions.schedule_c.models import ScheduleC
 from fecfiler.transactions.schedule_c.views import save_hook
 from fecfiler.transactions.models import Transaction
 from fecfiler.reports.models import Report
-from fecfiler.committee_accounts.views import create_committee_view
+from fecfiler.committee_accounts.utils import create_committee_view
 
 
 class ScheduleCViewsTestCase(TestCase):
