+++ django-backend/fecfiler/transactions/tests/test_models.py
@@ -1,7 +1,7 @@
 from decimal import Decimal
 from django.test import TestCase
 from fecfiler.reports.tests.utils import create_form3x
-from fecfiler.committee_accounts.views import create_committee_view
+from fecfiler.committee_accounts.utils import create_committee_view
 from fecfiler.committee_accounts.models import CommitteeAccount
 from fecfiler.transactions.models import Transaction
 from fecfiler.contacts.tests.utils import (
