+++ django-backend/fecfiler/web_services/dot_fec/tests/dot_fec_record_testing/test_schedule_c.py
@@ -3,7 +3,7 @@ from fecfiler.web_services.dot_fec.dot_fec_composer import compose_transactions
 from fecfiler.web_services.dot_fec.dot_fec_serializer import serialize_instance
 from fecfiler.web_services.dot_fec.dot_fec_serializer import FS_STR
 from fecfiler.committee_accounts.models import CommitteeAccount
-from fecfiler.committee_accounts.views import create_committee_view
+from fecfiler.committee_accounts.utils import create_committee_view
 from fecfiler.reports.tests.utils import create_form3x
 from fecfiler.transactions.tests.utils import create_loan
 from fecfiler.contacts.models import Contact
