+++ django-backend/fecfiler/web_services/dot_fec/tests/test_dot_fec_composer.py
@@ -1,13 +1,13 @@
 from django.test import TestCase
 from fecfiler.web_services.dot_fec.dot_fec_composer import (
     compose_dot_fec,
-    add_row_to_content
+    add_row_to_content,
 )
-from fecfiler.committee_accounts.views import create_committee_view
+from fecfiler.committee_accounts.utils import create_committee_view
 from fecfiler.web_services.dot_fec.dot_fec_serializer import (
     serialize_instance,
     CRLF_STR,
-    FS_STR
+    FS_STR,
 )
 from fecfiler.committee_accounts.models import CommitteeAccount
 from fecfiler.reports.tests.utils import create_form3x, create_form99, create_report_memo
@@ -115,7 +115,7 @@ class DotFECSerializerTestCase(TestCase):
         split_content = content.split("\n")
         split_report_row = split_content[1].split(FS_STR)
         self.assertEqual(split_report_row[14], "ABC\r")
-        free_text = content[content.find("[BEGINTEXT]"):]
+        free_text = content[content.find("[BEGINTEXT]") :]
         self.assertEqual(
             free_text,
             "[BEGINTEXT]\r\n\nBEHOLD! A large text string"
