+++ django-backend/fecfiler/web_services/summary/test_summary.py
@@ -1,7 +1,7 @@
 from decimal import Decimal
 from django.test import TestCase
 from fecfiler.committee_accounts.models import CommitteeAccount
-from fecfiler.committee_accounts.views import create_committee_view
+from fecfiler.committee_accounts.utils import create_committee_view
 from .summary import SummaryService
 from fecfiler.reports.tests.utils import create_form3x
 from datetime import datetime
